+++ hyperspy/_components/skew_normal.py
@@ -18,10 +18,12 @@
 
 import numpy as np
 import dask.array as da
+import sympy
 
 from hyperspy._components.expression import Expression
+from hyperspy.misc.utils import is_binned # remove in v2.0
+
 from distutils.version import LooseVersion
-import sympy
 
 sqrt2pi = np.sqrt(2 * np.pi)
 
@@ -211,7 +213,9 @@ class SkewNormal(Expression):
             self.A.value = height * sqrt2pi
             self.scale.value = scale
             self.shape.value = shape
-            if self.binned:
+            if is_binned(signal) is True:
+            # in v2 replace by
+            #if axis.is_binned:
                 self.A.value /= axis.scale
             return True
         else:
@@ -219,7 +223,9 @@ class SkewNormal(Expression):
                 self._create_arrays()
             self.A.map['values'][:] = height * sqrt2pi
 
-            if self.binned:
+            if is_binned(signal) is True:
+            # in v2 replace by
+            #if axis.is_binned:
                 self.A.map['values'] /= axis.scale
             self.A.map['is_set'][:] = True
             self.x0.map['values'][:] = x0