+++ hyperspy/_components/voigt.py
@@ -21,6 +21,7 @@ import sympy
 
 from hyperspy._components.expression import Expression
 from hyperspy._components.gaussian import _estimate_gaussian_parameters
+from hyperspy.misc.utils import is_binned # remove in v2.0
 from distutils.version import LooseVersion
 
 sqrt2pi = math.sqrt(2 * math.pi)
@@ -167,14 +168,18 @@ class Voigt(Expression):
             self.centre.value = centre
             self.sigma.value = sigma
             self.area.value = height * sigma * sqrt2pi
-            if self.binned:
+            if is_binned(signal) is True:
+            # in v2 replace by
+            #if axis.is_binned:
                 self.area.value /= axis.scale
             return True
         else:
             if self.area.map is None:
                 self._create_arrays()
             self.area.map['values'][:] = height * sigma * sqrt2pi
-            if self.binned:
+            if is_binned(signal) is True:
+            # in v2 replace by
+            #if axis.is_binned:
                 self.area.map['values'][:] /= axis.scale
             self.area.map['is_set'][:] = True
             self.sigma.map['values'][:] = sigma