+++ hyperspy/tests/io/test_tiff.py
@@ -388,8 +388,7 @@ FEI_Helios_metadata = {'Acquisition_instrument': {'SEM': {'Stage': {'rotation': 
                                    'authors': 'supervisor',
                                    'date': '2016-06-13',
                                    'time': '17:06:40'},
-                       'Signal': {'binned': False,
-                                  'signal_type': ''},
+                       'Signal': {'signal_type': ''},
                        '_HyperSpy': {'Folding': {'original_axes_manager': None,
                                                  'original_shape': None,
                                                  'signal_unfolded': False,
@@ -441,7 +440,7 @@ def test_read_Zeiss_SEM_scale_metadata_1k_image():
                       'original_filename': 'test_tiff_Zeiss_SEM_1k.tif',
                       'time': '09:40:32',
                       'title': ''},
-          'Signal': {'binned': False, 'signal_type': ''},
+          'Signal': {'signal_type': ''},
           '_HyperSpy': {'Folding': {'original_axes_manager': None,
                                     'original_shape': None,
                                     'signal_unfolded': False,
@@ -472,7 +471,7 @@ def test_read_Zeiss_SEM_scale_metadata_512_image():
                       'original_filename': 'test_tiff_Zeiss_SEM_512pix.tif',
                       'time': '08:20:42',
                       'title': ''},
-          'Signal': {'binned': False, 'signal_type': ''},
+          'Signal': {'signal_type': ''},
           '_HyperSpy': {'Folding': {'original_axes_manager': None,
                                     'original_shape': None,
                                     'signal_unfolded': False,
@@ -565,7 +564,7 @@ def test_read_TVIPS_metadata():
           'General': {'original_filename': 'TVIPS_bin4.tif',
                       'time': '9:01:17',
                       'title': ''},
-          'Signal': {'binned': False, 'signal_type': ''},
+          'Signal': {'signal_type': ''},
           '_HyperSpy': {'Folding': {'original_axes_manager': None,
                                     'original_shape': None,
                                     'signal_unfolded': False,
@@ -592,17 +591,20 @@ def test_axes_metadata():
         s2 = hs.load(fname)
         assert s2.axes_manager.navigation_axes[0].name == 'image series'
         assert s2.axes_manager.navigation_axes[0].units == nav_unit
+        assert s2.axes_manager.navigation_axes[0].is_binned == False
 
         fname2 = os.path.join(tmpdir, 'axes_metadata_IYX.tif')
         s.save(fname2, metadata={'axes':'IYX'})
         s3 = hs.load(fname2)
         assert s3.axes_manager.navigation_axes[0].name == 'image series'
         assert s3.axes_manager.navigation_axes[0].units == nav_unit
+        assert s3.axes_manager.navigation_axes[0].is_binned == False
 
         fname2 = os.path.join(tmpdir, 'axes_metadata_ZYX.tif')
         s.save(fname2, metadata={'axes':'ZYX'})
         s3 = hs.load(fname2)
         assert s3.axes_manager.navigation_axes[0].units == nav_unit
+        assert s3.axes_manager.navigation_axes[0].is_binned == False
 
 
 def test_olympus_SIS():
@@ -641,3 +643,4 @@ def test_save_angstrom_units():
             assert s2.axes_manager[0].units == s.axes_manager[0].units
         assert s2.axes_manager[0].scale == s.axes_manager[0].scale
         assert s2.axes_manager[0].offset == s.axes_manager[0].offset
+        assert s2.axes_manager[0].is_binned == s.axes_manager[0].is_binned